Change interface
Delete company interface
Join the user interfaces
