--- src/interface/db/CompanyInterface.ts
@@ -1,38 +1,0 @@
-export interface Company {
-  id_company?: number;
-  name: string;
-  address: string;
-  password: string;
-  rating: number;
-  photo: string;
-  phone: string;
-  web_site: string;
-}
-
-export interface Branch {
-  id: string;
-  name: string;
-
-  // Foreign Key
-  idCompany: string;
-}
-
-export interface Job {
-  id: string;
-  name: string;
-  state: boolean;
-  salary: string;
-  description: string;
-  requirements: string;
-
-  // Foreign Key
-  idBranch: string;
-}
-
-export interface Application {
-  state: 'accepted' | 'pending' | 'rejected';
-
-  // Foreign Key
-  userId: string;
-  companyId: string;
-}
