+++ src/interface/db/UserInterface.ts
@@ -22,3 +22,24 @@ export interface Document {
   name: 'curriculum' | 'aditional';
   route: string;
 }
+
+
+export interface Job {
+  id: string;
+  name: string;
+  state: boolean;
+  salary: string;
+  description: string;
+  requirements: string;
+
+  // Foreign Key
+  idBranch: string;
+}
+
+export interface Offer {
+  state: 'accepted' | 'pending' | 'rejected';
+
+  // Foreign Key
+  userId: string;
+  companyId: string;
+}
